migrate prismo to flakes; deprecate secrets.nix
deprecate secrets.nix for the moment due to nix flake rebuild taking an
independent copy of the GH checkout and not using the local checkout w/ tmpfiles.

add flake.nix which points (currently in version control) to prismo.nix

TODO: point this to a symlink `configuration.nix` which points then to the local
deployment.